FIX: make the dataset doctest fixture modular
