+++ .gitignore
@@ -12,6 +12,7 @@ dist/
 doc/_build/
 doc/auto_examples/
 doc/modules/generated/
+doc/datasets/generated/
 pip-log.txt
 scikits.learn.egg-info/
 .coverage
